Consolidate marshaler interfaces
