+++ entity.go
@@ -2,6 +2,7 @@ package estoria
 
 import (
 	"context"
+	"encoding/json"
 
 	"github.com/go-estoria/estoria/typeid"
 )
@@ -14,10 +15,40 @@ type Entity interface {
 	ApplyEvent(ctx context.Context, eventData EntityEvent) error
 }
 
+type EntityMarshaler[E Entity] interface {
+	Marshal(entity E) ([]byte, error)
+	Unmarshal(data []byte, dest *E) error
+}
+
+type JSONEntityMarshaler[E Entity] struct{}
+
+func (JSONEntityMarshaler[E]) Marshal(entity E) ([]byte, error) {
+	return json.Marshal(entity)
+}
+
+func (JSONEntityMarshaler[E]) Unmarshal(data []byte, dest *E) error {
+	return json.Unmarshal(data, dest)
+}
+
 // EntityEvent is an event that can be applied to an entity to change its state.
 type EntityEvent interface {
 	EventType() string
 	New() EntityEvent
+}
+
+type EntityEventMarshaler interface {
+	Unmarshal(b []byte, d EntityEvent) error
+	Marshal(d EntityEvent) ([]byte, error)
+}
+
+type JSONEntityEventMarshaler struct{}
+
+func (s JSONEntityEventMarshaler) Unmarshal(b []byte, d EntityEvent) error {
+	return json.Unmarshal(b, d)
+}
+
+func (s JSONEntityEventMarshaler) Marshal(d EntityEvent) ([]byte, error) {
+	return json.Marshal(d)
 }
 
 // A DiffableEntity is an entity that can be diffed against another entity
